refactor: 예약 정보 패키지 분리
