+++ src/main/java/roomescape/controller/reservation/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
