+++ src/main/java/roomescape/controller/reservation/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.reservation;
 
 public record ReservationResponse(Long id, String name, String date, String time) {
 }
